nieuwe saml, oude user -> update user
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWOProject/DWO-LMS/DWOServer/trunk@23597 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/test/resources/fi/dwo/server/mysql/InitTestDatabase.sql
@@ -468,7 +468,7 @@ CREATE TABLE `tblsamluser` (
   `samluserid` varchar(255) NOT NULL,
   `userID` int(11) NOT NULL,
   `authtoken` varchar(16) NOT NULL DEFAULT '97358736458',
-  `timestampauthtoken` int(11) NOT NULL DEFAULT '0',
+  `timestampauthtoken` bigint(20) NOT NULL DEFAULT '0',
   `optlock` int(11) DEFAULT '0',
   `lastChangeTimeStamp` bigint(20) DEFAULT '0',
   `del` tinyint(4) NOT NULL DEFAULT '0',
